Rasterizing
I think the algorithm is applied per-pixel, so I have rasterized the circle and pulled out the points normalised to the global coordinate system.
Next step is to feed these coordinates into the formula, see what happens and perform convex hull filtering to get back to a vector-based path again.